# update deriv_smooth_vel and implicitfast