1. generalize the implementation of `deriv_smooth_vel` for dof trees and actuator configurations
2. update `implicitfast` integrator to support `sparse`

update `deriv_smooth_vel`
- general actuator configurations
- tendon damping
- sparse

update `implicitfast`
- integrator `implicit` supports sparse
- inverse dynamics `implicitfast` supports sparse

testing:
- position actuators with ordering that is not grouped by dof tree
- tendon damping
- implicit integrator sparse
- inverse dynamics sparse

**(benchmarks from before adding support for `sparse`)**

there is some amount of performance regression with the updated implementation:

```
mjwarp-testspeed --function=step --mjcf=test_data/humanoid/humanoid.xml --batch_size=8192 --nconmax=512000 --njmax=512000 --integrator="implicitfast" --event_trace
```

this pr:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.29 s
 Total simulation time: 34.01 s
 Total steps per second: 240,879
 Total realtime factor: 1,204.39 x
 Total time per step: 4151.47 ns

Event trace:

step: 4147.38
  forward: 3993.46
    fwd_position: 294.92
      kinematics: 39.78
      com_pos: 15.64
      camlight: 3.79
      crb: 57.17
      tendon_armature: 0.17
      factor_m: 29.96
      collision: 37.70
        nxn_broadphase: 6.71
        convex_narrowphase: 0.18
        primitive_narrowphase: 29.30
      make_constraint: 104.85
      transmission: 4.12
    sensor_pos: 0.17
    fwd_velocity: 60.12
      com_vel: 18.69
      passive: 2.55
      rne: 25.25
      tendon_bias: 0.18
    sensor_vel: 0.18
    fwd_actuation: 15.11
    fwd_acceleration: 25.65
      xfrc_accumulate: 8.70
      solve_m: 15.17
    sensor_acc: 0.17
    solve: 3595.32
      mul_m: 16.90
      _solver_iteration: [ 768.57, 672.91, 329.69, 185.72, 167.69, 161.09, 159.98, 159.85, 159.76, 159.62 ]
        _linesearch: [ 142.84, 117.09, 54.81, 37.43, 32.81, 31.03, 30.81, 30.76, 30.74, 30.71 ]
          mul_m: [ 16.99, 15.59, 5.99, 3.77, 2.70, 2.26, 2.21, 2.19, 2.19, 2.19 ]
  implicit: 153.40
    deriv_smooth_vel: 120.29
```

main:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.29 s
 Total simulation time: 33.02 s
 Total steps per second: 248,075
 Total realtime factor: 1,240.38 x
 Total time per step: 4031.03 ns

Event trace:

step: 4026.95
  forward: 3976.93
    fwd_position: 294.01
      kinematics: 39.62
      com_pos: 15.60
      camlight: 3.77
      crb: 57.13
      tendon_armature: 0.17
      factor_m: 29.85
      collision: 37.51
        nxn_broadphase: 6.67
        convex_narrowphase: 0.17
        primitive_narrowphase: 29.16
      make_constraint: 104.52
      transmission: 4.11
    sensor_pos: 0.17
    fwd_velocity: 59.94
      com_vel: 18.63
      passive: 2.54
      rne: 25.19
      tendon_bias: 0.17
    sensor_vel: 0.17
    fwd_actuation: 15.07
    fwd_acceleration: 25.56
      xfrc_accumulate: 8.67
      solve_m: 15.12
    sensor_acc: 0.17
    solve: 3580.01
      mul_m: 16.85
      _solver_iteration: [ 766.45, 670.12, 328.03, 184.52, 166.63, 160.03, 158.94, 158.84, 158.77, 158.66 ]
        _linesearch: [ 142.48, 116.68, 54.55, 37.14, 32.57, 30.79, 30.56, 30.52, 30.51, 30.49 ]
          mul_m: [ 16.95, 15.54, 5.96, 3.75, 2.69, 2.24, 2.19, 2.17, 2.18, 2.17 ]
  implicit: 49.50
    deriv_smooth_vel: 16.52
```