@thowell 2025-07-10
i think one of the benefits of the existing implementation is that it processes groups of actuators with tile operations. 

an assumption the current implementation makes is that actuators are grouped according to their dof tree. the new test mjcf `actuation/position.xml` provides an example case where the actuator ordering is not grouped by dof tree. the updated implementation now processes actuators separately.

this pr aims to generalize `deriv_smooth_vel` and test additional environments. in a follow-up pr i think it is probably possible to improve the performance of the updated `deriv_smooth_vel` implementation.