@adenzler-nvidia 2025-07-11
Let's focus on getting the behavior right first, I think this does not really count as a regression. Thanks a lot for fixing this. Maybe there's also ways of getting back this perf without tiles. Clearly my assumptions were wrong here.